ensure all participants are matched exactly once (#29)
Fixes #28

This particular issue required 4 criteria to be met for it it to possibly occur

1. The grouping size has to be more than 2
2. There are already at least 2 participants chosen for a specific match
3. There are historical matches from a previous run
4. The chosen participants within a particular match have no candidates in common

When all of that is met, it was possible for someone to be selected again because the code wasn't filtering down to only the unmatched remaining participants. It may sound like an edge case, but it really is more of a late stage bug within the matching logic, so it was possible to happen every time.