Paginate through slack channels
If there are over 1000 slack channels in an organization, which is
entirely possible with the ability to archive channels, the call method
in loads_slack_channels would only retrieve the first 1000.

This change adds a method to follow any pagination cursors and combine
channel list pages together to ensure we have a complete channel list.

This change also makes use of the exclude_archived argument for
conversations_list which will (on Slack's side) throw out any archived
channels. A snippet to keep in mind from Slack's docs:

https://api.slack.com/methods/conversations.list
"When paginating, any filters used in the request are applied after
retrieving a virtual page's limit. For example. using
exclude_archived=true when limit=20 on a virtual page that would contain
15 archived channels will return you the virtual page with only 5
results. Additional results are available from the next cursor
valuePaginate through slack channels."